Safari-611.1.10/LayoutTests/fast/scrolling/mac/non-invertible-transform-hit-testing-expected.txt


Test scroll over content
PASS overflowScrollEventCount is 0
PASS windowScrollEventCount > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
